Crate zng_wgt_checkerboard

Checkerboard widget, properties and nodes.

Crate

This crate is part of the zng project.

Structs

Checkerboard

W A checkerboard visual.

Colors

Checker board colors.

Statics

COLORS_VAR

The checkerboard colors.

ORIGIN_VAR

Offset applied to the checkerboard pattern.

SIZE_VAR

The size of one color rectangle in the checkerboard.

Functions

cb_origin

P Sets the offset of the checkerboard pattern.

cb_size

P Set the size of a checkerboard color rectangle.

colors

P Set both checkerboard colors.

node

Checkerboard node.