Struct zng_wgt_checkerboard::Checkerboard

pub struct Checkerboard(/* private fields */);

W A checkerboard visual.

This widget draws a checkerboard pattern, with configurable dimensions and colors.

Implementations

impl Checkerboard

pub fn widget_new() -> Self

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl Checkerboard

pub fn colors(&self, colors: impl IntoVar<Colors>)

P Set both checkerboard colors.

The values are the interchanging colors for a given color scheme, for example in the dark color scheme the (colors[0].dark, colors[1].dark) colors are used.

This property sets COLORS_VAR for all inner checkerboard widgets.

impl Checkerboard

pub fn cb_size(&self, size: impl IntoVar<Size>)

P Set the size of a checkerboard color rectangle.

This property sets the SIZE_VAR for all inner checkerboard widgets.

impl Checkerboard

pub fn cb_origin(&self, offset: impl IntoVar<Point>)

P Sets the offset of the checkerboard pattern.

Relative values are resolved in the context of a cb_size.

This property sets the ORIGIN_VAR for all inner checkerboard widgets.